Replace definition of variable sum from int to unsigned long
 The maximum size of integer variable on 64bit machines is 2^31. Suppose the
 writer takes 5 nanosecond to increase the counter by one, the result wraps
 around in about 10 seconds.

+++ CodeSamples/count/count_stat_eventual.c
@@ -38,7 +38,7 @@ unsigned long read_count(void)
 void *eventual(void *arg)
 {
 	int t;
-	int sum;
+	unsigned long sum;
 
 	while (stopflag < 3) {
 		sum = 0;
